feedlist move to feed screen
